requirements.txt
ccxt==4.3.89
pandas==2.2.2
ta==0.11.0
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
